Added 'EventFlag::Auto'
Tries to guess what type of connection to use between direct and queued. When it is impossible to make a guess, direct connection is used.
Note: wx objects do not store a reference to the thread they belong. If two objects live on separate thread, it is required to explicitly use queued connection.