Added CallbackCommand::Discard enum member
Supposed use: if the callback caller implementation requires, this value can be used to notify it that the callback didn't not process the data (skipped or discarded it), but still needs to be called again for any new data (no termination is going to happen). The caller is not supposed to increment call counters or perform any other post-processing in case of CallbackCommand::Discard as opposed to CallbackCommand::Continue.